refactor(Heading): remove unnecessary component
