--- source/components/atoms/Heading/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Heading";
