+++ source/components/atoms/index.ts
@@ -1,7 +1,6 @@
 export { default as Box } from "./Box";
 export { default as Button } from "./Button";
 export { default as Fieldset } from "./Fieldset";
-export { default as Heading } from "./Heading";
 export { default as Icon } from "./Icon";
 export { default as Input } from "./Input";
 export { default as Text } from "./Text";
