+++ source/components/molecules/StoryWrapper.tsx
@@ -1,10 +1,11 @@
 import PropTypes from "prop-types";
 import React from "react";
 import styled from "styled-components/native";
-import Heading from "../atoms/Heading";
+
+import { Text } from "../atoms";
 import ScreenWrapper from "./ScreenWrapper";
 
-const StoryHeading = styled(Heading)`
+const StoryHeading = styled(Text).attrs({ type: "h2" })`
   margin-bottom: 16px;
   margin-top: 16px;
 `;
@@ -20,9 +21,7 @@ const StoryWrapper = (props) => {
   return (
     <EnhancedSafeAreaView>
       <ScreenWrapper style={style}>
-        {kind ? (
-          <StoryHeading type="h2">{`${kind} / ${name}`}</StoryHeading>
-        ) : null}
+        {kind ? <StoryHeading>{`${kind} / ${name}`}</StoryHeading> : null}
         {children}
       </ScreenWrapper>
     </EnhancedSafeAreaView>
